+++ iotbay/pom.xml
@@ -5,11 +5,11 @@
   <modelVersion>4.0.0</modelVersion>
 
   <groupId>com.uts</groupId>
-  <artifactId>iotbay</artifactId>
+  <artifactId>labs</artifactId>
   <version>1.4</version>
   <packaging>war</packaging>
 
-  <name>iotbay Maven Webapp</name>
+  <name>labs Maven Webapp</name>
   <!-- FIXME change it to the project's website -->
   <url>http://www.example.com</url>
 
@@ -19,9 +19,6 @@
     <maven.compiler.target>1.8</maven.compiler.target>
   </properties>
 
-
-
-
   <!-- DEPENDENCIES -->
   <dependencies>
     <dependency>
@@ -30,22 +27,29 @@
       <version>4.11</version>
       <scope>test</scope>
     </dependency>
+    <dependency>
+        <groupId>org.postgresql</groupId>
+        <artifactId>postgresql</artifactId>
+        <version>42.7.5</version>
+    </dependency>
     <!-- https://mvnrepository.com/artifact/org.xerial/sqlite-jdbc -->
     <dependency>
-      <groupId>org.xerial</groupId>
-      <artifactId>sqlite-jdbc</artifactId>
-      <version>3.49.1.0</version>
+        <groupId>org.xerial</groupId>
+        <artifactId>sqlite-jdbc</artifactId>
+        <version>3.49.1.0</version>
+    </dependency>
+    <!-- https://mvnrepository.com/artifact/jakarta.servlet/jakarta.servlet-api -->
+    <dependency>
+        <groupId>jakarta.servlet</groupId>
+        <artifactId>jakarta.servlet-api</artifactId>
+        <version>6.1.0</version>
+        <scope>provided</scope>
     </dependency>
   </dependencies>
 
+
   <build>
-    <resources>
-      <resource>
-        <directory>src/main/java</directory>
-      </resource>
-    </resources>
-    <directory>src/main/webapp/WEB-INF</directory>
-    <finalName>iotbay</finalName>
+    <finalName>labs</finalName>
     <pluginManagement><!-- lock down plugins versions to avoid using Maven defaults (may be moved to parent pom) -->
       <plugins>
         <plugin>
