--- iotbay/src/main/webapp/WEB-INF/classes/model/dao/DB.java
@@ -1,13 +1,0 @@
-package model.dao;
-
-import java.sql.Connection;
-
-/**
- * Stores database information
- */
-
-public abstract class DB {
-    protected String URL = "jdbc:sqlite:iotbay/database/iotbay.db";
-    protected String driver = "org.sqlite.JDBC";
-    protected Connection conn;
-}
