--- iotbay/src/main/webapp/WEB-INF/classes/model/dao/DBConnector.java
@@ -1,32 +1,0 @@
-
-package model.dao;
-
-import java.sql.Connection;
-
-import java.sql.DriverManager;
-
-import java.sql.SQLException;
-
-public class DBConnector extends DB {
-
-    public DBConnector() throws ClassNotFoundException, SQLException {
-
-        Class.forName(driver);
-
-        conn = DriverManager.getConnection(URL);
-
-    }
-
-    public Connection openConnection() {
-
-        return this.conn;
-
-    }
-
-    public void closeConnection() throws SQLException {
-
-        this.conn.close();
-
-    }
-
-}
